# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/chat/msg": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "发送消息",
                "parameters": [
                    {
                        "description": "消息请求体",
                        "name": "messageReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.MessageReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            }
        },
...
        "/api/contact/getContactList": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取联系人列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "游标，用于分页查询",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认为20",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/contact/getNewContactList": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取新的联系人列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "时间戳，用于获取此时间之后的新联系人",
                        "name": "timestamp",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/contact/getNewMsgList": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取新的消息列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "消息ID",
                        "name": "msgId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "房间ID",
                        "name": "roomId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/contact/userInfo/batch": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取批量用户信息",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "用户ID列表",
                        "name": "uidList",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            }
        },
...
        "/api/group/create": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "创建群聊",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "群聊成员ID列表",
                        "name": "uidList",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "req.MessageBody": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "replyMsgId": {
                    "type": "integer"
                }
            }
        },
        "req.MessageReq": {
            "type": "object",
            "required": [
                "body",
                "msgType",
                "roomId"
            ],
            "properties": {
                "body": {
                    "$ref": "#/definitions/req.MessageBody"
                },
                "msgType": {
                    "type": "integer"
                },
                "roomId": {
                    "type": "integer"
                }
            }
        },
        "resp.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                },
                "success": {
                    "description": "是否成功",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}